Refactor configure commnad.
